requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
sentence-transformers
openai
elevenlabs
langdetect
pandas
openpyxl
httpx
slowapi
python-dotenv